// lib/api/branch/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type BranchInfo = {
	/**
	 * The remote associated with the branch.
	 */
	remote?: string | null;
	/**
	 * The name of the branch.
	 */
	name?: string;
	/**
	 * Indicates whether the branch is local or remote.
	 */
	kind?: "Local" | "Remote";
};

/**
 * Indicates whether the branch is local or remote.
 */
export type kind = "Local" | "Remote";

export type GetRepoBranchByRepoPathData = {
	repoPath?: string;
};

export type GetRepoBranchByRepoPathResponse = Array<BranchInfo>;

export type PostRepoBranchByRepoPathData = {
	name?: string;
	repoPath?: string;
};

export type PostRepoBranchByRepoPathResponse = unknown;

export type DeleteRepoBranchByRepoPathData = {
	name?: string;
	repoPath?: string;
};

export type DeleteRepoBranchByRepoPathResponse = unknown;

export type PutRepoBranchByRepoPathData = {
	repoPath?: string;
	to?: string;
};

export type PutRepoBranchByRepoPathResponse = unknown;